test(use-local-storage): add browser-mode hook tests
